refactor: move TransactionBatch into separate module
